WeaponSystems

Documentation for WeaponSystems.

Introduction

This package creates used to describe weapon systems

julia> using WeaponSystems

Table of contents

WeaponSystems

Introduction

Table of contents

Package Features

Package Features

Targets

Circular and rectangular target definition

Weapon systems

Tank

WeaponSystems.Canon — Type

The Canon is defined by:
* θ elevation angle (degrees)
* lw the length of the canon
* u₀ the muzzle velocity magnitude

WeaponSystems.Hull — Type

The Hull is defined by :
* Φ: angle of rotation of the tank around the longitudinal axis. If the right side of the tank is lower than the left side,
than this angle is positif.
* Ψ: angle defined by the longitudinal axes of the tank and the horizontal plane
If the nose of the tank is higher than the back, this angle is positif
* Χ: angle defined by the the projection of the longitudinal axis of the tank into an horizontal plane
and the X axis

WeaponSystems.Sight — Type

The sight is defined by:
* θ LOS elevation angle
* ξ LOS azimuth angle

WeaponSystems.Tank — Type

The tank is defined by: a hull, a turret, a canon, a sight and its position (altitude)

WeaponSystems.TargetCirc — Type

TargetCirc(ρ,position)

defines a circular target by specifying the radius ρ and the position (range)

WeaponSystems.TargetRect — Type

TargetRect(a,b,position)

defines a rectangular target by specifying the lenght a, the height b and the position (range)

WeaponSystems.TargetSphe — Type

TargetSphe(ρ,position)

defines a spherical target by specifying the radius ρ and the position (range)

WeaponSystems.Turret — Type

The turret is defined by its azimuth angle

WeaponSystems.muzzlePos — Method

muzzlePos(tank)

Returns the coordinates of the muzzle exit in a coordinate system which has the origin on the tank at the base of the the canon

WeaponSystems.targetPos — Method

targetPos(target,tank)

Is a function that returns the coordinates of the target in a coordinate system which has the origin on the tank at the base of the the canon

WeaponSystems.Canon

WeaponSystems.Hull

WeaponSystems.Sight

WeaponSystems.Tank

WeaponSystems.TargetCirc

WeaponSystems.TargetRect

WeaponSystems.TargetSphe

WeaponSystems.Turret

WeaponSystems.muzzlePos

WeaponSystems.targetPos