WeaponSystems

Documentation for WeaponSystems.

WeaponSystems.TargetCirc

WeaponSystems.TargetRect

WeaponSystems.Turret

WeaponSystems.TargetCirc — Type

TargetCirc(ρ,position)

defines a circular target by specifying the radius ρ and the position (range)

WeaponSystems.TargetRect — Type

TargetRect(a,b,position)

defines a rectangular target by specifying the lenght a, the height b and the position (range)

WeaponSystems.Turret — Type

The turret is defined by its azimuth angle